Add a context menu and missing capability to the text editor #7549
