+++ plugins/transforms/json/src/main/java/org/apache/hop/pipeline/transforms/types/JsonExplorerFileType.java
@@ -48,6 +48,8 @@ public class JsonExplorerFileType extends BaseTextExplorerFileType<JsonExplorerF
             IHopFileType.CAPABILITY_CLOSE,
             IHopFileType.CAPABILITY_FILE_HISTORY,
             IHopFileType.CAPABILITY_COPY,
+            IHopFileType.CAPABILITY_CUT,
+            IHopFileType.CAPABILITY_PASTE,
             IHopFileType.CAPABILITY_SELECT));
   }
