+++ plugins/transforms/textfile/src/main/java/org/apache/hop/pipeline/transforms/types/MarkDownExplorerFileTypeHandler.java
@@ -23,17 +23,16 @@ import java.io.FileOutputStream;
 import java.io.OutputStream;
 import java.nio.charset.StandardCharsets;
 import java.util.List;
-import org.apache.hop.core.Props;
 import org.apache.hop.core.gui.plugin.GuiPlugin;
 import org.apache.hop.core.gui.plugin.toolbar.GuiToolbarElement;
+import org.apache.hop.core.gui.plugin.toolbar.GuiToolbarElementFilter;
 import org.apache.hop.core.gui.plugin.toolbar.GuiToolbarElementType;
+import org.apache.hop.ui.core.FormDataBuilder;
 import org.apache.hop.ui.core.PropsUi;
 import org.apache.hop.ui.core.dialog.ErrorDialog;
-import org.apache.hop.ui.core.gui.GuiToolbarWidgets;
-import org.apache.hop.ui.core.gui.IToolbarContainer;
+import org.apache.hop.ui.core.widget.editor.IContentEditorWidget;
 import org.apache.hop.ui.hopgui.ContentEditorFacade;
 import org.apache.hop.ui.hopgui.HopGui;
-import org.apache.hop.ui.hopgui.ToolbarFacade;
 import org.apache.hop.ui.hopgui.perspective.explorer.ExplorerFile;
 import org.apache.hop.ui.hopgui.perspective.explorer.ExplorerPerspective;
 import org.apache.hop.ui.hopgui.perspective.explorer.file.types.text.BaseTextExplorerFileTypeHandler;
@@ -45,19 +44,16 @@ import org.commonmark.ext.task.list.items.TaskListItemsExtension;
 import org.commonmark.node.Node;
 import org.commonmark.parser.Parser;
 import org.commonmark.renderer.html.HtmlRenderer;
-import org.eclipse.swt.SWT;
-import org.eclipse.swt.layout.FormAttachment;
-import org.eclipse.swt.layout.FormData;
 import org.eclipse.swt.widgets.Composite;
-import org.eclipse.swt.widgets.Control;
 
 /** How do we handle a markdown file in the file explorer perspective? */
-@GuiPlugin
+@GuiPlugin(name = "Markdown file type handler")
 public class MarkDownExplorerFileTypeHandler extends BaseTextExplorerFileTypeHandler {
 
   private static final Class<?> PKG = MarkDownExplorerFileType.class;
 
-  public static final String TOOLBAR_PARENT_ID = "MarkDownExplorerFileTypeHandler-Toolbar";
+  protected static final String LANGUAGE = "markdown";
+
   public static final String TOOLBAR_ITEM_PREVIEW =
       "MarkDownExplorerFileTypeHandler-ToolBar-Preview";
 
@@ -72,41 +68,17 @@ public class MarkDownExplorerFileTypeHandler extends BaseTextExplorerFileTypeHan
 
   @Override
   protected String getLanguageId() {
-    return "markdown";
+    return LANGUAGE;
   }
 
   @Override
   public void renderFile(Composite composite) {
-    // Create the toolbar container using the facade
-    IToolbarContainer toolBarContainer =
-        ToolbarFacade.createToolbarContainer(composite, SWT.WRAP | SWT.LEFT | SWT.HORIZONTAL);
-    Control toolBarControl = toolBarContainer.getControl();
-
-    // Create the toolbar widgets using GuiToolbarWidgets
-    GuiToolbarWidgets toolBarWidgets = new GuiToolbarWidgets();
-    toolBarWidgets.registerGuiPluginObject(this);
-    toolBarWidgets.createToolbarWidgets(toolBarContainer, TOOLBAR_PARENT_ID);
-
-    // Layout data for the toolbar control
-    FormData fdToolBar = new FormData();
-    fdToolBar.left = new FormAttachment(0, 0);
-    fdToolBar.right = new FormAttachment(100, 0);
-    fdToolBar.top = new FormAttachment(0, 0);
-    toolBarControl.setLayoutData(fdToolBar);
-    toolBarControl.pack();
-    PropsUi.setLook(toolBarControl, Props.WIDGET_STYLE_TOOLBAR);
 
     // Create the standard editor widget
     editorWidget = ContentEditorFacade.createContentEditor(composite, getLanguageId());
 
     // Position the editor widget below the toolbar
-    Control editorControl = editorWidget.getControl();
-    FormData fdEditor = new FormData();
-    fdEditor.left = new FormAttachment(0, 0);
-    fdEditor.right = new FormAttachment(100, 0);
-    fdEditor.top = new FormAttachment(toolBarControl, 0);
-    fdEditor.bottom = new FormAttachment(100, 0);
-    editorControl.setLayoutData(fdEditor);
+    editorWidget.getControl().setLayoutData(FormDataBuilder.builder().fullSize().build());
 
     // If it's a new file, there's no need to reload it
     if (this.getFilename() != null) {
@@ -123,14 +95,26 @@ public class MarkDownExplorerFileTypeHandler extends BaseTextExplorerFileTypeHan
         });
   }
 
+  @GuiToolbarElementFilter(parentId = IContentEditorWidget.GUI_PLUGIN_TOOLBAR_PARENT_ID)
+  public static boolean showForMarkdownFileType(String itemId, Object guiPluginInstance) {
+    if (TOOLBAR_ITEM_PREVIEW.equals(itemId)
+        && guiPluginInstance instanceof IContentEditorWidget editor) {
+      return LANGUAGE.equals(editor.getLanguage());
+    }
+    return true;
+  }
+
   @GuiToolbarElement(
-      root = TOOLBAR_PARENT_ID,
+      root = IContentEditorWidget.GUI_PLUGIN_TOOLBAR_PARENT_ID,
       id = TOOLBAR_ITEM_PREVIEW,
       toolTip = "i18n::MarkDownFileTypeHandler.Preview.Tooltip",
       type = GuiToolbarElementType.BUTTON,
-      image = "ui/images/preview.svg")
+      image = "ui/images/preview.svg",
+      separator = true)
+  //  public static void previewMarkdown(IContentEditorWidget editorWidget) {
   public void previewMarkdown() {
     try {
+      // IContentEditorWidget editorWidget = (IContentEditorWidget) event.widget;
       String markdown = editorWidget.getText();
 
       // Parse markdown to HTML body content
@@ -304,7 +288,10 @@ public class MarkDownExplorerFileTypeHandler extends BaseTextExplorerFileTypeHan
       EnvironmentUtils.getInstance().openUrl(tempFile.toURI().toString());
     } catch (Exception e) {
       new ErrorDialog(
-          hopGui.getShell(), "Error", "Error generating or displaying Markdown preview", e);
+          HopGui.getInstance().getActiveShell(),
+          "Error",
+          "Error generating or displaying Markdown preview",
+          e);
     }
   }
 }
