+++ plugins/transforms/textfile/src/main/java/org/apache/hop/pipeline/transforms/types/ShellExplorerFileType.java
@@ -48,6 +48,8 @@ public class ShellExplorerFileType extends BaseTextExplorerFileType<TextExplorer
             IHopFileType.CAPABILITY_CLOSE,
             IHopFileType.CAPABILITY_FILE_HISTORY,
             IHopFileType.CAPABILITY_COPY,
+            IHopFileType.CAPABILITY_CUT,
+            IHopFileType.CAPABILITY_PASTE,
             IHopFileType.CAPABILITY_SELECT));
   }
