+++ plugins/transforms/textfile/src/main/java/org/apache/hop/pipeline/transforms/types/TextExplorerFileType.java
@@ -48,6 +48,8 @@ public class TextExplorerFileType extends BaseTextExplorerFileType<TextExplorerF
             IHopFileType.CAPABILITY_CLOSE,
             IHopFileType.CAPABILITY_FILE_HISTORY,
             IHopFileType.CAPABILITY_COPY,
+            IHopFileType.CAPABILITY_CUT,
+            IHopFileType.CAPABILITY_PASTE,
             IHopFileType.CAPABILITY_SELECT));
   }
