+++ ui/src/main/resources/org/apache/hop/ui/hopgui/messages/messages_en_US.properties
@@ -327,3 +327,25 @@ HopGui.LayoutCheck.Dialog.ReplaceDummyTitle=Replace Transform
 HopGui.LayoutCheck.Dialog.ReplaceDummyMessage=Do you want to replace this Dummy transform with a ''Stream Schema Merge'' transform?
 HopGui.LayoutCheck.Dialog.StreamSchemaPluginNotFound=Stream Schema Merge plugin not found.
 HopGui.LayoutCheck.Dialog.ReplaceDummyError=Error replacing Dummy transform with Stream Schema Merge
+ContentEditorWidget.Menu.Copy=Copy
+ContentEditorWidget.Menu.Cut=Cut
+ContentEditorWidget.Menu.Paste=Paste
+ContentEditorWidget.Menu.Redo=Redo
+ContentEditorWidget.Menu.SelectAll=Select All
+ContentEditorWidget.Menu.Undo=Undo
+ContentEditorWidget.Menu.Copy=Copy
+ContentEditorWidget.Menu.Cut=Cut
+ContentEditorWidget.Menu.Paste=Paste
+ContentEditorWidget.Menu.Redo=Redo
+ContentEditorWidget.Menu.SelectAll=Select All
+ContentEditorWidget.Menu.UnselectAll=Clear selection
+ContentEditorWidget.Menu.Undo=Undo
+ContentEditorWidget.ToolBar.SelectAll.Tooltip=Select all text
+ContentEditorWidget.ToolBar.UnselectAll.Tooltip=Clear selection
+ContentEditorWidget.ToolBar.Copy.Tooltip=Copy selected text to clipboard
+ContentEditorWidget.ToolBar.Cut.Tooltip=Cut selected text to clipboard
+ContentEditorWidget.ToolBar.Paste.Tooltip=Paste text from clipboard
+ContentEditorWidget.ToolBar.Undo.Tooltip=Undo the last action
+ContentEditorWidget.ToolBar.Redo.Tooltip=Redo the last action
+ContentEditorWidget.ToolBar.Find.Tooltip=Find text
+ContentEditorWidget.ToolBar.Replace.Tooltip=Replace text
